move hooks to separated file
